Small bugfix to micrometer implementation
