+++ agroal-spring-boot-starter/src/main/java/io/agroal/springframework/boot/metrics/AgroalPoolDataSourceMetadataProviderConfiguration.java
@@ -4,18 +4,11 @@
 package io.agroal.springframework.boot.metrics;
 
 import io.agroal.api.AgroalDataSource;
-import io.agroal.pool.DefaultMetricsRepository;
-import io.micrometer.core.instrument.MeterRegistry;
-import io.micrometer.core.instrument.binder.MeterBinder;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.jdbc.DataSourceUnwrapper;
 import org.springframework.boot.jdbc.metadata.DataSourcePoolMetadataProvider;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-
-import java.util.Map;
-
-import javax.sql.DataSource;
 
 @Configuration(proxyBeanMethods = false)
 @ConditionalOnClass(AgroalDataSource.class)
@@ -30,32 +23,5 @@ public class AgroalPoolDataSourceMetadataProviderConfiguration {
             }
             return null;
         };
-    }
-
-    @Bean
-    @ConditionalOnClass(MeterRegistry.class)
-    AgroalDataSourceMeterBinder agroalDataSourceMeterBinder(Map<String, DataSource> dataSources) {
-        return new AgroalDataSourceMeterBinder(dataSources);
-    }
-
-    static class AgroalDataSourceMeterBinder implements MeterBinder {
-
-        private final Map<String, DataSource> dataSources;
-
-        AgroalDataSourceMeterBinder(Map<String, DataSource> dataSources) {
-            this.dataSources = dataSources;
-        }
-
-        @Override
-        public void bindTo(MeterRegistry registry) {
-            dataSources.forEach((name, dataSource) -> bindDataSourceToRegistry(name, dataSource, registry));
-        }
-
-        private void bindDataSourceToRegistry(String name, DataSource dataSource, MeterRegistry registry) {
-            AgroalDataSource agroalDataSource = DataSourceUnwrapper.unwrap(dataSource, AgroalDataSource.class);
-            if (agroalDataSource != null && agroalDataSource.getMetrics() instanceof DefaultMetricsRepository) {
-                new AgroalDataSourcePoolMetrics(name, agroalDataSource).bindTo(registry);
-            }
-        }
     }
 }
